+++ pkg/math/geom/circle.go
@@ -5,12 +5,12 @@ import (
 )
 
 type Circle struct {
-	X float64
-	Y float64
-	R float64
+	X float32
+	Y float32
+	R float32
 }
 
-func MakeCircle(x, y, r float64) Circle {
+func MakeCircle(x, y, r float32) Circle {
 	return Circle{
 		X: x,
 		Y: y,
@@ -18,15 +18,15 @@ func MakeCircle(x, y, r float64) Circle {
 	}
 }
 
-func (c Circle) Diameter() float64 {
+func (c Circle) Diameter() float32 {
 	return 2 * c.R
 }
 
 func (c Circle) HasOverlap(other Circle) bool {
-	return math.Pow((c.X-other.X), 2)+math.Pow((c.Y-other.Y), 2) <=
-		math.Pow(2*(c.R+other.R), 2)
+	return math.Pow(float64(c.X-other.X), 2)+math.Pow(float64(c.Y-other.Y), 2) <=
+		math.Pow(2*float64(c.R+other.R), 2)
 }
 
 func (c Circle) ContainsPoint(p Point) bool {
-	return math.Abs((c.X-p.X)*(c.X-p.X)+(c.Y-p.Y)*(c.Y-p.Y)) < ((c.Diameter()) * (c.Diameter()))
+	return math.Abs(float64((c.X-p.X)*(c.X-p.X)+(c.Y-p.Y)*(c.Y-p.Y))) < float64((c.Diameter())*(c.Diameter()))
 }
